% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_response.R
\name{irf_penalized}
\alias{irf_penalized}
\title{IRFs for penalized tvvar_fit (factor-uncertainty only)}
\usage{
irf_penalized(
  fit,
  lags = 10,
  T.max = 5,
  B = 500,
  shock_index = 1,
  shock_position = NULL,
  seed = 1234
)
}
\arguments{
\item{fit}{tvvar_fit from penalized_estimate() or unpenalized_estimate()}

\item{lags}{horizon (number of periods ahead)}

\item{T.max}{number of time points (from start) at which to compute IRFs}

\item{B}{Monte Carlo draws (factor simulations only)}

\item{shock_index}{index of shocked variable (1..N) \link{for convenience when using ek as unit vector}}

\item{shock_position}{numeric length-N shock vector (if given, overrides shock_index)}

\item{seed}{base RNG seed}
}
\value{
list(IRF_lb, IRF_med, IRF_ub)
}
\description{
IRFs for penalized tvvar_fit (factor-uncertainty only)
}
