% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_lib_package_building.R
\name{lavar_fit}
\alias{lavar_fit}
\title{fit a Lasso_VAR model based on LS with l1 penalty}
\usage{
lavar_fit(it, X, lam, b, alpha, A_old, A_oldold)
}
\arguments{
\item{it}{time}

\item{X}{T*d array which stores the data of the time series. T is the length and d is the dimension.}

\item{lam}{shrinkage parameter}

\item{b}{bandwidth}

\item{alpha}{stepsize}

\item{A_old}{initial matrix 1}

\item{A_oldold}{initial matrix 2}
}
\value{
Estimation of transition matrix
}
\description{
fit a Lasso_VAR model based on LS with l1 penalty
}

