% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_lib_package_building.R
\name{lsvar_fit}
\alias{lsvar_fit}
\title{fit a Ridge_VAR model based on LS with l2 penalty}
\usage{
lsvar_fit(it, X, lam, b)
}
\arguments{
\item{it}{time}

\item{X}{T*d array which stores the data of the time series. T is the length and d is the dimension.}

\item{lam}{penalty parameter}

\item{b}{bandwidth}
}
\value{
Estimation of transition matrix
}
\description{
fit a Ridge_VAR model based on LS with l2 penalty
}

