% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_lib_package_building.R
\name{mlevar_fit}
\alias{mlevar_fit}
\title{fit a MLE_VAR model based on MLE}
\usage{
mlevar_fit(it, X, b)
}
\arguments{
\item{it}{time}

\item{X}{T*d array which stores the data of the time series. T is the length and d is the dimension.}

\item{b}{bandwidth}
}
\value{
Estimation of transition matrix
}
\description{
fit a MLE_VAR model based on MLE
}

