% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvirf.R
\name{plot.tvirf}
\alias{plot.tvirf}
\title{Plot IRFs (single or multiple shocks)}
\usage{
\method{plot}{tvirf}(x, t = 1, ...)
}
\arguments{
\item{x}{a \code{tvirf()} object (single IRF list or multi-shock list)}

\item{t}{which time index to plot (default 1)}

\item{...}{Additional graphical arguments passed to lower-level plotting functions.}
}
\description{
Plot IRFs (single or multiple shocks)
}
