% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_lib_package_building.R
\name{ridvar_fit}
\alias{ridvar_fit}
\title{Time-varying Ridge model}
\usage{
ridvar_fit(it, X, lam, b)
}
\arguments{
\item{it}{Specify the time point we are researching}

\item{X}{T*d array which stores the data of the time series. T is the length of time and d is the dimension.}

\item{lam}{penalty parameter}

\item{b}{bandwidth}
}
\value{
Estimation of transition matrix at time it
}
\description{
Fit a time-varying Ridge model for vector autoregressive data
}

