% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated TV-VAR dataset}
\format{
A list with:
\describe{
\item{data}{T x N matrix of simulated observations.}
\item{Phi.f.array}{Array of factor loadings used.}
\item{u}{Innovations.}
\item{factors}{Latent factor series.}
}
}
\usage{
simdata
}
\description{
Artificial dataset generated from a TV-VAR(1) model with 2 variables
and 1 factor. The parameters used in the simulation are:
}
\details{
\describe{
\item{dim.VAR}{2}
\item{lag.order}{1}
\item{number.factors}{1}
\item{T}{1000 observations}
\item{A2}{0.1}
\item{B2}{0.75}
\item{pphi}{0.95}
\item{omega}{\eqn{\begin{bmatrix}0.3 & 0.2 \\ 0.2 & 0.3\end{bmatrix}}}
\item{Phi.c}{\eqn{\begin{bmatrix}0.3 & 0.15 \\ 0.0 & 0.3\end{bmatrix}}}
\item{Phi.f}{\eqn{\begin{bmatrix}0.2 & 0.0 \\ 0.0 & 0.2\end{bmatrix}}}
}
}
\keyword{datasets}
