% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvsim.R
\name{tvsim}
\alias{tvsim}
\title{Simulate data from a TV-VAR with scalar BEKK innovations}
\usage{
tvsim(
  T = 200,
  N = 2,
  p = 1,
  r = 1,
  A2 = 0.1,
  B2 = 0.75,
  pphi = 0.95,
  omega = matrix(c(0.3, 0.2, 0.2, 0.3), 2),
  Phi.c = matrix(c(0.3, 0.15, 0, 0.3), 2),
  Phi.f = matrix(c(0.2, 0, 0, 0.2), 2),
  burn_in = 500
)
}
\arguments{
\item{T}{Number of time points (after burn-in).}

\item{N}{Number of variables.}

\item{p}{Lag order of the VAR.}

\item{r}{Number of factors.}

\item{A2}{Scalar BEKK parameter A.}

\item{B2}{Scalar BEKK parameter B.}

\item{pphi}{If r=1: scalar. If r>1: (rxr) matrix with pphi values on the diagonal.}

\item{omega}{Covariance matrix (N x N).}

\item{Phi.c}{Static VAR coefficients (N x (N*p)).}

\item{Phi.f}{Factor loadings matrix (N x (N*pr)).}

\item{burn_in}{Number of burn-in steps (default 500).}
}
\value{
List with simulated data matrix Y (T x N), latent factors, shocks, and Phi.f structure.
}
\description{
Simulate data from a TV-VAR with scalar BEKK innovations
}
