% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvvar-package.R
\docType{package}
\name{tvvar-package}
\alias{tvvar}
\alias{tvvar-package}
\title{tvvar: Time-Varying VAR Models}
\description{
GS: This file contains the package documentation and globals if any
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gijssmeets/tvvar}
  \item Report bugs at \url{https://github.com/gijssmeets/tvvar/issues}
}

}
\author{
\strong{Maintainer}: Julia Schaumburg \email{j.schaumburg@vu.nl} (\href{https://orcid.org/0000-0003-1832-0355}{ORCID})

Authors:
\itemize{
  \item Lennart Hoogerheide \email{l.f.hoogerheide@vu.nl} (\href{https://orcid.org/0000-0002-3318-1202}{ORCID})
  \item Gijs Smeets \email{gijsolinsmeets@gmail.com}
}

}
\keyword{internal}
