% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_lib_package_building.R
\name{lavar_fit}
\alias{lavar_fit}
\title{Time-varying Lasso model}
\usage{
lavar_fit(it, X, lam, b, alpha, A_old, A_oldold)
}
\arguments{
\item{it}{Specify the time point we are researching}

\item{X}{T*d array which stores the data of the time series. T is the length of time and d is the dimension.}

\item{lam}{shrinkage parameter}

\item{b}{bandwidth}

\item{alpha}{stepsize}

\item{A_old}{initial matrix 1}

\item{A_oldold}{initial matrix 2}
}
\value{
Estimation of transition matrix at time it
}
\description{
Fit a time-varying Lasso model for vector autoregressive data
}

