% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized.R
\name{penalized_estimate}
\alias{penalized_estimate}
\title{Penalized TV-VAR estimation via ECM with adaptive L1 weights}
\usage{
penalized_estimate(
  data,
  p,
  r,
  phi_f_structure,
  zero_mean = TRUE,
  lambda = c(0.015, 0.015),
  control = list(tol = 0.001, max_iter = 200, trace = TRUE, maxit = 2000)
)
}
\arguments{
\item{data}{T x N numeric matrix}

\item{p}{lag order}

\item{r}{number of factors}

\item{phi_f_structure}{restrictions: 3D \link{N,N,r}, list of r (N x N), or N x (N*r) block}

\item{zero_mean}{logical; TRUE fixes intercept to 0}

\item{lambda}{length-2 numeric: c(lambda_c, lambda_f)}

\item{control}{list(tol, max_iter, trace, maxit for inner ML if needed)}
}
\description{
Penalized TV-VAR estimation via ECM with adaptive L1 weights
}
