% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvirf.R
\name{plot.tvirf}
\alias{plot.tvirf}
\title{Plot IRFs (single or multiple shocks)}
\usage{
\method{plot}{tvirf}(irf_obj, t = 1)
}
\arguments{
\item{irf_obj}{result of irf() (single IRF list or multi-shock list)}

\item{t}{which time index to plot (default 1)}
}
\description{
Plot IRFs (single or multiple shocks)
}
