% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_response.R
\name{plot_irf}
\alias{plot_irf}
\title{Plot IRFs (single or multiple shocks)}
\usage{
plot_irf(irf_obj, t = 1)
}
\arguments{
\item{irf_obj}{result of irf()}

\item{t}{which time index (default 1)}
}
\description{
Plot IRFs (single or multiple shocks)
}
