% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpenalized.R
\name{unpenalized_estimate}
\alias{unpenalized_estimate}
\title{Unpenalized estimation (ML or EM) for TV-VAR}
\usage{
unpenalized_estimate(
  data,
  p,
  r,
  zero_mean = TRUE,
  phi_f_structure,
  method = c("ML", "EM"),
  control = list(maxit = 2000, trace = 0),
  em_control = list(max_iter = 200, tol = 0.001, trace = TRUE)
)
}
\description{
Unpenalized estimation (ML or EM) for TV-VAR
}
